+++ src/Response/Model/FeedItem.php
@@ -10,24 +10,28 @@ use InstagramAPI\AutoPropertyMapper;
  * @method Ad4ad getAd4ad()
  * @method int getAdLinkType()
  * @method mixed getEndOfFeedDemarcator()
+ * @method mixed getExploreStory()
  * @method Item getMediaOrAd()
  * @method StoriesNetego getStoriesNetego()
  * @method SuggestedUsers getSuggestedUsers()
  * @method bool isAd4ad()
  * @method bool isAdLinkType()
  * @method bool isEndOfFeedDemarcator()
+ * @method bool isExploreStory()
  * @method bool isMediaOrAd()
  * @method bool isStoriesNetego()
  * @method bool isSuggestedUsers()
  * @method $this setAd4ad(Ad4ad $value)
  * @method $this setAdLinkType(int $value)
  * @method $this setEndOfFeedDemarcator(mixed $value)
+ * @method $this setExploreStory(mixed $value)
  * @method $this setMediaOrAd(Item $value)
  * @method $this setStoriesNetego(StoriesNetego $value)
  * @method $this setSuggestedUsers(SuggestedUsers $value)
  * @method $this unsetAd4ad()
  * @method $this unsetAdLinkType()
  * @method $this unsetEndOfFeedDemarcator()
+ * @method $this unsetExploreStory()
  * @method $this unsetMediaOrAd()
  * @method $this unsetStoriesNetego()
  * @method $this unsetSuggestedUsers()
@@ -40,6 +44,7 @@ class FeedItem extends AutoPropertyMapper
         'ad4ad'                  => 'Ad4ad',
         'suggested_users'        => 'SuggestedUsers',
         'end_of_feed_demarcator' => '',
+        'explore_story'          => '',
         'ad_link_type'           => 'int',
     ];
 }